Refactor: Layer separation
